Change _run() to accept a list not a string
Accept commands as a list like ["ln", "-s" ... ] (the format that
subprocess.check_output() uses) instead of as a string like "ln -s ... ".

Construct the rsync, mv, rm and ln commands in snapshot() as lists not strings.

Also change the rm and ln commands to support running over ssh as the mv
command does.

Update tests.