Begin to implement removing old snapshots
The algorithm is: Run rsync, if it fails with "no space left on device" then
remove the oldest snapshot, repeat until rsync succeeds or fails for some
other reason. Raises an error if there are less than --min-snapshots snapshots
remaining.

Needs testing on local and remote filesystems that are running out of
space.

Needs unit tests.